requirements.txt
termcolor
pyyaml
setuptools
checksumdir
wget
gdown
google-api-python-client
google-auth-httplib2
google-auth-oauthlib